add null checks
