+++ src/Microsoft.AspNetCore.SystemWebAdapters/Adapters/RemoteAppClientExtensions.cs
@@ -1,18 +1,18 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System;
 using Microsoft.AspNetCore.SystemWebAdapters;
-using Microsoft.Extensions.DependencyInjection;
 
-namespace System.Web;
+namespace Microsoft.Extensions.DependencyInjection;
 
-public static class RemoteAppExtensions
+public static class RemoteAppClientExtensions
 {
     /// <summary>
     /// Add configuration for connecting to a remote app for System.Web extensions that require a remote
     /// ASP.NET app such as remote app authentication or remote app session sharing.
     /// </summary>
-    public static ISystemWebAdapterBuilder AddRemoteAppServer(this ISystemWebAdapterBuilder builder, Action<ISystemWebAdapterRemoteAppBuilder> configure)
+    public static ISystemWebAdapterBuilder AddRemoteAppClient(this ISystemWebAdapterBuilder builder, Action<ISystemWebAdapterRemoteAppBuilder> configure)
     {
         if (builder is null)
         {
@@ -24,9 +24,8 @@ public static class RemoteAppExtensions
             throw new ArgumentNullException(nameof(configure));
         }
 
-        var options = builder.Services.AddOptions<RemoteAppOptions>()
-            .Validate(options => !string.IsNullOrEmpty(options.ApiKey), "ApiKey must be set")
-            .Validate(options => !string.IsNullOrEmpty(options.ApiKeyHeader), "ApiKeyHeader must be set");
+        builder.Services.AddOptions<RemoteAppOptions>()
+            .ValidateDataAnnotations();
 
         configure(new Builder(builder.Services));
 
@@ -35,6 +34,16 @@ public static class RemoteAppExtensions
 
     public static ISystemWebAdapterRemoteAppBuilder Configure(this ISystemWebAdapterRemoteAppBuilder builder, Action<RemoteAppOptions> configure)
     {
+        if (builder is null)
+        {
+            throw new ArgumentNullException(nameof(builder));
+        }
+
+        if (configure is null)
+        {
+            throw new ArgumentNullException(nameof(configure));
+        }
+
         builder.Services.AddOptions<RemoteAppOptions>()
             .Configure(configure);
